Added option to plot source energy, and plot top n bottom n
Added sourceEnergy boolean variable, to allow the function to plot electricity
and gas in source energy form, i.e. a multiplier of 3.14 and 1.05 are applied to
electricity and gas.

topn and botn are for specifying which subset of buildings in the "score" list
to include in the regional report: best performing 4 and worst performing 16 are
the default.