Corrected state and city name, added auto-backup to write table
Added column state_abbr to be the state name to be used, as the original "State"
column has a mixture of full name and abbreviation.

Unified city names, from a external csv file created by me

Implemented an auto-backup feature in write_table_to_db, so that a backup csv of
the table being overwritten will be created automatically in "db_build_temp_csv/backups".