Resolved duplicate in output of left_join state_abbr
In creating the state full name to state abbreviation lookup table, I attached
state abbreviation to map to itself to the bottom of the lookup table, so that
I can map a combination of full name and abbreviation to abbreviation. However
there is a duplicate of the state abbreviation column in the source table
state_abbr (DC), so I got two copies of rows for DC buildings. I removed the
duplicate when appending the state abbreviation -> state abbreviation, and
resolved the duplicates in the output.